Adding some hello-world functional tests
 * Also changed the asset cache file to be CWD agnostic
