+++ tests/functional/fixtures/flags/bar_on.json
@@ -1,0 +1,3 @@
+{
+  "feature_flags.bar": true
+}
