+++ tests/functional/ui/__init__.py
@@ -1,0 +1,3 @@
+class TestLandingPage:
+    def test_we_can_get_the_page(self):
+        pass
