+++ tox.ini
@@ -45,15 +45,16 @@ passenv =
 deps =
     {tests,clean,coverage}: coverage
     {tests,lint}: httpretty
-    {tests,lint,docstrings,checkdocstrings}: pytest
-    {tests,lint,docstrings,checkdocstrings}: factory-boy
+    {tests,functests,lint,docstrings,checkdocstrings}: pytest
+    {tests,functests,lint,docstrings,checkdocstrings}: factory-boy
+    {functests,lint,docstrings,checkdocstrings}: webtest
     lint: prospector==1.1.6.2
     {format,checkformatting}: black
     {format,checkformatting}: isort
     {docstrings,checkdocstrings}: sphinx
     docstrings: sphinx-autobuild
     docker-compose: docker-compose
-    {tests,lint,docstrings,checkdocstrings}: -r requirements.txt
+    {tests,functests,lint,docstrings,checkdocstrings}: -r requirements.txt
     dev: -r requirements-dev.in
     pip-compile: pip-tools
 setenv =
@@ -64,6 +65,7 @@ whitelist_externals =
 commands =
     tests: sh bin/create-testdb
     tests: coverage run -m pytest -v -Werror {posargs:tests/unit/}
+    functests: pytest {posargs:tests/functional/}
     dev: {posargs:gunicorn --paste conf/development.ini}
     lint: prospector
     lint: prospector tests
